chore: refactor OnPointer<Pressed|Released>
